[PP-132] Added Merge all Collector class to EngineDataCollector class
