+++ Project-Pentagram/Engine/Collector/ShaderCollector.cpp
@@ -1,0 +1,15 @@
+#include "ShaderCollector.h"
+
+ShaderCollector::ShaderCollector()
+	// Load each shader of object
+	: GameObjectShader("Shader/GameObjVertex.shader","Shader/GameObjFragment.shader"),
+	  TextShader("Shader/TextVertex.shader", "Shader/TextFragment.shader")
+{
+	
+}
+
+
+void ShaderCollector::Free()
+{
+
+}
