+++ Project-Pentagram/Engine/Collector/ShaderCollector.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include "../Renderer/Shader.h"
+
+class ShaderCollector
+{
+	public:
+		// Shader for use with object
+		Shader GameObjectShader;
+		Shader TextShader;
+
+		ShaderCollector();
+		void Free();
+};
