+++ Project-Pentagram/Engine/Collector/TextureCollector.h
@@ -1,0 +1,17 @@
+#pragma once
+
+#include <assert.h>
+#include <string>
+#include <map>
+
+#include "../Renderer/Texture.h"
+
+class TextureCollector
+{
+	private:
+		std::map<std::string, Texture*> textures;
+	public:
+		TextureCollector();
+		Texture* GetTexture(const std::string& name);
+		void Free();
+};
