+++ Project-Pentagram/Engine/Scene/MenuScene.h
@@ -3,7 +3,7 @@
 #include "../GameScene.h"
 #include "../SceneManager.h"
 #include "../Input.h"
-#include "../Renderer/ShaderCollector.h"
+#include "../Collector/EngineDataCollector.h"
 
 class MenuScene : public GameScene
 {
